refactor: Habit data is formatted client-side
